requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.1.0
opencv-python>=4.8.0
face-recognition>=1.3.0
dlib>=19.24.0
Pillow>=10.0.0
python-dateutil>=2.8.2